Move payment checkout to its own object
