+++ lib/catarse_paypal_express.rb
@@ -3,6 +3,8 @@ require "catarse_paypal_express/engine"
 require "catarse_paypal_express/gateway"
 require "catarse_paypal_express/contribution_actions"
 require "catarse_paypal_express/payment"
+require "catarse_paypal_express/payment_setup"
+require "catarse_paypal_express/payment_checkout"
 require "catarse_paypal_express/transaction_fee_calculator_base"
 require "catarse_paypal_express/transaction_additional_fee_calculator"
 require "catarse_paypal_express/transaction_inclusive_fee_calculator"
